vim: Add textobj-user, textobj-comment
This helps with the annoying problem of trying to reformat Ruby comments
that are adjacent to code. Generally what happens by default is all of
the lines are joined and it does not work at all. Instead, now I can say
`gqaC` to reformat just the lines that are in the comment. This helps
with preventing comment lines from getting too long.